Check whether material src attribute is string or element
In aframe <= v0.3, the src attribute for a material was a string, with
a DOM selector in it. for aframe >= v0.3, the src attribute is the
actual DOM element for the src, saving a step for the developer.
Fixing this makes the rest of this plugin work again. Tested in aframe
v0.3.2 and aframe v0.5.0.

Fixes #14
Fixes #16